cleaned sequence api
